Implement AA pattern for C++ target
Note that special care has to be taken to provide ghost cells around
active cells so the algorithm has somewhere to stream to and from.

This is also the case for the AB pattern but there they only have to
be equilibrilized once instead of after every other time step.

Even when such an equilibrilization is performed there is still a
potential bug as inbound populations at the outer boundary are never
streamed to (this is not a problem for AB using pull-only streaming).
A vectorizable solution may require direction-specific ghost cell
equilibrization.