Implement basic version of the SSS pattern for C++ target
An interesting extension of the AA pattern.  The main advantage of this is
that updating pointers in a control structure is  much more elegant than
duplicating all function implementations as is required by the normal AA
pattern. For more details see [1].

Only works for the SOA layout.

On a pure memory access level this pattern is equivalent to the AA pattern.
The difference is how the memory locations are calculated (by pointer swap
& shift or by different indexing functions for odd and even time steps).

[1]: "An auto-vectorization friendly parallel lattice Boltzmann streaming
      scheme for direct addressing" by Mohrhard et al. (2019)