Implement SSS for OpenCL target
Sadly OpenCL kernels don't accept pointer-to-pointer arguments which
complicates the control structure implementation.

A workaround is to cast them into `uintptr_t` which is guaranteed to be large
enough to fit any pointer on the device. Special care has to be taken to always
perform the pointer shifts on actual floating point pointers and not on
type-less pointers.